% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_matrix_list.R
\docType{data}
\name{a_matrix_list}
\alias{a_matrix_list}
\title{A sample sf_network of Oxford region}
\format{
the data is a list
\describe{
   \item{a_matrix_list[[1]]}{Origin-Destination pairs pertaining to the shortest route output in List[[4]]}
   \item{a_matrix_list[[2]]}{Origin nodes pertaining to the shortest route output in List[[4]]}
   \item{a_matrix_list[[3]]}{Destination nodes pertaining to the shortest route output in List[[4]]}
   \item{a_matrix_list[[4]]}{Numeric vector (of 0s and 1s) indication the shortest routes}

}
}
\usage{
data(a_matrix_list) #Lazy loading since this will not be loaded onto the memory.
The data becomes available only when the user queries it.
}
\description{
This is an sf_network of Oxford region where the lines and polygon data have been spatially joined.
This data provides an example of what a sf_network objects looks like and also is an argument to run many of the example
codes.
}
\keyword{datasets}
