% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_CensusOD_points.R
\name{add_CensusOD_points}
\alias{add_CensusOD_points}
\title{Adds Origin destination points on the network}
\usage{
add_CensusOD_points(region, crs, sf_network)
}
\arguments{
\item{region}{The PCT region or local authority to download data from (e.g. oxford or Leeds). See View(pct_regions_lookup) for a full list of possible region names.}

\item{crs}{coordinate reference system}

\item{sf_network}{should be of class: "sfnetwork" "tbl_graph" "igraph"}
}
\description{
Adds orgin destination points derived from the library pct (propensity to cycle tool) on to the nearest nodes on the
sf_network provided.
}
\examples{
osm_data<-extract_data( c( -1.41785, -1.09785, 51.59201, 51.91201)) # for Oxford region
osm_roads_data<-modify_Linesdata(osm_data, road_types = c('motorway', 'motorway_link', 'trunk','trunk_link', 'primary', 'primary_link', 'secondary','secondary_link' ,'tertiary' , 'tertiary_link' ) ,road_levels = NULL,crs=9766)
osm_polygon_data<-modify_Polygondata(osm_data,road_types=c('motorway', 'motorway_link', 'trunk','trunk_link', 'primary', 'primary_link', 'secondary','secondary_link' ,'tertiary' , 'tertiary_link' ), road_levels = NULL,crs=9766 )
osm_roads_processed<-lineString_rounding(osm_roads_data,rounding=3) \%>\% as_sfnetwork(directed= FALSE) \%>\% simplify_sf_data()
osm_polygons_processed<-lineString_rounding(osm_polygon_data,rounding=3) \%>\% as_sfnetwork(directed= FALSE) \%>\% simplify_sf_data()
joined_script <- st_network_join(osm_roads_processed, osm_polygons_processed) \%>\% network_preProcessing()
}
