% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addODnodes.R
\name{addODnodes}
\alias{addODnodes}
\title{Map origin destination nodes on the sf_network}
\usage{
addODnodes(lat_col, long_col, crs, sf_network_directed, inputCoordinatedf)
}
\arguments{
\item{inputCoordinatedf}{}
}
\description{
Map origin destination nodes on the sf_network
}
