% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_direction.R
\name{assign_direction}
\alias{assign_direction}
\title{Determines the direction of the bearings}
\usage{
assign_direction(bearings, userDataSliced)
}
\arguments{
\item{userDataSliced}{}
}
\description{
Determines the direction of the bearings
}
