% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_initial_bearing.R
\name{calculate_initial_bearing}
\alias{calculate_initial_bearing}
\title{Calculate bearing}
\usage{
calculate_initial_bearing(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{- latitude of starting node}

\item{lon1}{- longitude of starting node}

\item{lat2}{- latitude of ending node}

\item{lon2}{- longitude of ending node}
}
\value{
Bearing in degrees
}
\description{
Calculate bearing
}
