% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_0_1.R
\name{convert_0_1}
\alias{convert_0_1}
\title{Producing matrix columns of 1s and 0s}
\usage{
convert_0_1(x, matrix_column_length)
}
\arguments{
\item{x}{Every vector in a list}

\item{matrix_column_length}{The length of every vector in argument x.}
}
\value{
A vector of 0s and 1s where 1s indicate the row index that is part of the shortest path and
0s indicate the route index not part of the shortest path.
}
\description{
This function will be called exclusively from create.Amatrix function.
}
