% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tmap.R
\name{draw_tmap}
\alias{draw_tmap}
\title{Draw map}
\usage{
draw_tmap(
  sf_network_directed,
  shortestPath_network = NULL,
  shortestPathReverse_network = NULL,
  ToRowNumber = NULL,
  FromRowNumber = NULL,
  col1 = "red",
  col2 = "pink",
  lwd = 3,
  lty = "dashed",
  shape = 8
)
}
\arguments{
\item{sf_network_directed}{should be of class: "sfnetwork" "tbl_graph" "igraph"}

\item{shortestPath_network}{should be of class: "sfnetwork" "tbl_graph" "igraph"}

\item{shortestPathReverse_network}{should be of class: "sfnetwork" "tbl_graph" "igraph"}

\item{ToRowNumber}{Origin node relating to the list of origin destination input data}

\item{FromRowNumber}{Destination node relating to the list of origin destination input data}

\item{col1}{colour for shortestPath. Default set to red.}

\item{col2}{colour for reverse_shortestPath. Default set to pink.}

\item{lwd}{Line width. Default set to 3.}

\item{lty}{Line type. Default set to "dashed".}

\item{shape}{Line shape. Default set to 8.}
}
\value{
Tmap object
}
\description{
Draw map
}
