% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CensusOD_data.R
\name{extract_CensusOD_data}
\alias{extract_CensusOD_data}
\title{Derives OD data from Census UK}
\usage{
extract_CensusOD_data(region, crs)
}
\arguments{
\item{region}{The PCT region or local authority to download data from (e.g. oxford or Leeds). See View(pct_regions_lookup) for a full list of possible region names.}

\item{crs}{coordinate reference system}
}
\value{
returnList[[1]] - contains traffic demands between any given ODs.
        returnList[[2]] - contains origin and destination nodes and their coordinates. This is an
        input to addODnodes() function.
}
\description{
Extracts the Census UK. The object returned is then passed onto addODnodes() function to map them on the
processed sf_network.
}
