% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Title Extract OSM data.}
\usage{
extract_data(bounding_coordinates)
}
\arguments{
\item{bounding_coordinates}{}
}
\value{
Data frame containing OSM data pertaining to the bouding coordinates
}
\description{
Title Extract OSM data.
}
\examples{
osm_data<-extract_data( c( -1.41785, -1.09785, 51.59201, 51.91201)) #for Oxford region
}
