% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEdgesDirection.R
\name{findEdgesDirection}
\alias{findEdgesDirection}
\title{Direction of the traffic flow information}
\usage{
findEdgesDirection(joined_script)
}
\arguments{
\item{joined_script}{}
}
\value{
joined_script with added column on traffic flow direction
}
\description{
Identifies direction of the traffic flow information.
Every edge within the network has traffic flow information and starting and ending coordinates. Using the start and end coordinates
this function calculates the cardinal and inter-cardinal direction.
}
\examples{
directiondf<-findEdgesDirection(joined_script)
}
