% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineString_rounding.R
\name{lineString_rounding}
\alias{lineString_rounding}
\title{Rounding off the co-coordinates}
\usage{
lineString_rounding(sfdata_dataframe, rounding = 3)
}
\arguments{
\item{rounding}{to which nearest decimal}
}
\value{
sf data frame
}
\description{
Rounding off the co-coordinates
}
\examples{
osm_data<-extract_data( c( -1.41785, -1.09785, 51.59201, 51.91201)) # for Oxford region
osm_roads_data<-modify_Linesdata(osm_data, road_types = c('motorway', 'motorway_link', 'trunk','trunk_link', 'primary', 'primary_link', 'secondary','secondary_link' ,'tertiary' , 'tertiary_link' ) ,road_levels = NULL,crs=9766)
osm_roads_data<-modify_Polygondata(osm_data, road_types = c('motorway', 'motorway_link', 'trunk','trunk_link', 'primary', 'primary_link', 'secondary','secondary_link' ,'tertiary' , 'tertiary_link' ) ,road_levels = NULL,crs=9766)
osm_roads_processed<-lineString_rounding(osm_roads_data,rounding=3)

}
