% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToEdges.R
\name{mapToEdges}
\alias{mapToEdges}
\title{Map user nodes onto sf_network edges}
\usage{
mapToEdges(userData, sf_network, crs, lat_col, long_col)
}
\arguments{
\item{long_col}{}
}
\description{
Map user nodes onto sf_network edges
}
