% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToEdges1.R
\name{mapToEdges1}
\alias{mapToEdges1}
\title{Map user nodes onto sf_network edges}
\usage{
mapToEdges1(userData, sf_network, crs, lat_col, long_col, SiteColname)
}
\arguments{
\item{long_col}{}
}
\description{
Map user nodes onto sf_network edges
}
