% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_Linesdata.R
\name{modify_Linesdata}
\alias{modify_Linesdata}
\title{Modifies the road/linear features of OSM data}
\usage{
modify_Linesdata(sf_data, road_types = NULL, road_levels = NULL, crs)
}
\arguments{
\item{sf_data}{OSM data extracted}

\item{road_types}{road types which are to be retained alternatively,}

\item{road_levels}{road level thresholds to be retained.}

\item{crs}{enter the appropriate coordinate reference system (crs) appropriate to the region of interest}
}
\value{
modified OSM linear/lines data containing user specified road types/levels - sf dataframe
}
\description{
Modifies the road/linear features of OSM data
}
\examples{
osm_data<-extract_data( c( -1.41785, -1.09785, 51.59201, 51.91201)) # for Oxford region
osm_roads_data<-modify_Linesdata(osm_data, road_types = c('motorway', 'motorway_link', 'trunk','trunk_link', 'primary', 'primary_link', 'secondary','secondary_link' ,'tertiary' , 'tertiary_link' ) ,road_levels = NULL,crs=9766)

}
