% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_Polygondata.R
\name{modify_Polygondata}
\alias{modify_Polygondata}
\title{Modifies the extracted polygon data from OSM}
\usage{
modify_Polygondata(sf_data, road_types, road_levels, crs)
}
\arguments{
\item{sf_data}{OSM data extracted}

\item{road_types}{road types which are to be retained alternatively,}

\item{road_levels}{road level thresholds to be retained.}

\item{crs}{enter the appropriate coordinate reference system (crs) appropriate to the region of interest}
}
\value{
modified OSM polygons data containing user specified road types/levels - sf data frame
}
\description{
Modifies the extracted polygon data from OSM
}
\examples{
osm_data<-extract_data( c( -1.41785, -1.09785, 51.59201, 51.91201)) # for Oxford region
osm_roads_data<-modify_Polygondata(osm_data, road_types = c('motorway', 'motorway_link', 'trunk','trunk_link', 'primary', 'primary_link', 'secondary','secondary_link' ,'tertiary' , 'tertiary_link' ) ,road_levels = NULL,crs=9766)

}
