% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_data.R
\docType{data}
\name{osm_data}
\alias{osm_data}
\title{Open street map data for Oxford region}
\format{
An object of class \code{list} (inherits from \code{osmdata}, \code{osmdata_sf}) of length 8.
}
\usage{
osm_data
}
\description{
The object returned is a list from OSM API.The list contains all geographic
street map information related to Oxford.
}
\keyword{datasets}
