% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbed_shortestPath.R
\name{perturbed_shortestPath}
\alias{perturbed_shortestPath}
\title{Title}
\usage{
perturbed_shortestPath(from, to, processed_roadNetwork, sd)
}
\arguments{
\item{from}{Origin node relating to the list of origin destination input data}

\item{to}{Destination node relating to the list of origin destination input data}

\item{sd}{Standard deviation is applied to re-calculate the time_travelled so to produce more than one shortest
routes between any given OD pairs.}
}
\value{
vector of row indices of edges part of the shortest path network
}
\description{
Title
}
