% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoadTransportModel.R
\name{RoadTransportModel}
\alias{RoadTransportModel}
\title{RoadTransportModel}
\usage{
RoadTransportModel(arg1, arg2 = NULL, arg3 = NULL, arg4)
}
\arguments{
\item{arg1}{Location coordinates}

\item{arg2}{Input road types or default to NULL}

\item{arg3}{Input road levels or default to NULL}

\item{arg4}{Coordinate reference system}
}
\description{
Executes all the steps in creating road transport model in  a single function. This includes line/node modification,
roundabout simplification, creating two way networks,  smoothing pseudo nodes etc.
}
