% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapped_geometry_script.R
\name{snapped_geometry_script}
\alias{snapped_geometry_script}
\title{Title}
\usage{
snapped_geometry_script(rowIndex, processed_roadNetwork, census_ODs)
}
\arguments{
\item{census_ODs}{}
}
\description{
Title
}
