% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewSf_network.R
\name{viewSf_network}
\alias{viewSf_network}
\title{View Sf network}
\usage{
viewSf_network(sf_network)
}
\arguments{
\item{sf_network}{}
}
\value{
map of the sf_network
}
\description{
View Sf network
}
\examples{
data(joined_script)
viewSf_network(joined_script)
}
