% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.Amatrix.R
\name{create_Amatrix}
\alias{create_Amatrix}
\title{Creates A-matrix}
\usage{
create_Amatrix(processed_roadNetwork, census_ODs, sd = 0.3)
}
\arguments{
\item{sd}{Standard deviation is applied to re-calculate the time_traveled so to produce more than one shortest
routes between any given OD pairs. Default set of 0.3.}
}
\value{
List. List[[1]]- Origin-Destination pairs pertaining to the shortest route output in List[[4]].
              List[[2]]-Origin nodes pertaining to the shortest route output in List[[4]].
              List[[3]]-Destination nodes pertaining to the shortest route output in List[[4]].
              List[[4]]- Numeric vector (of 0s and 1s) indication the shortest routes.
}
\description{
Shortest route is calculated using Dijkstra's algorithm using the weight time traveled (road speed/ length of road).
Potential shortest routes are also identified using perturbed time traveled where the weights are perturbed by a
standard deviation of 0.3 (user has the option to change sd per their choice).
Thus every origin destination  pair of nodes will have at least one shortest route and may have more than one
shortest routes based on the perturbed weight column (i.e., time traveled).
}
\examples{
data(joined_script)
data(censusOD_points)
aMatrix<-create_Amatrix(joined_script, censusOD_points[[2]])

}
