% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaData.trafficCounts.R
\name{metaData_trafficCounts}
\alias{metaData_trafficCounts}
\title{Title}
\usage{
metaData_trafficCounts(a_matrix_list, censusOD_points, region)
}
\arguments{
\item{region}{}
}
\description{
Title
}
