% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_pathInNetwork.R
\name{shortest_pathInNetwork}
\alias{shortest_pathInNetwork}
\title{Shortest Path in the Network}
\usage{
shortest_pathInNetwork(
  processed_roadNetwork,
  ToRowNumber,
  FromRowNumber,
  census_ODs
)
}
\arguments{
\item{processed_roadNetwork}{sf_network after undergoing network processing}

\item{census_ODs}{}
}
\value{
vector of row indices of edges part of the shortest path network
}
\description{
Shortest Path in the Network
}
